// src/app/features/gastos/application/dto/paginated-response.dto.ts
export interface PaginatedResponseDto<T> {
  code: number;
  status: string;
  message: string;
  data: T[];
  meta: {
    current_page: number;
    last_page: number;
    per_page: number;
    total: number;
  };
  links: {
    first: string;
    last: string;
    prev: string | null;
    next: string | null;
  };
}